+++ Changelog.md
@@ -1,3 +1,5 @@
 # Changelog
 - Improve generator compatibility when loading removed celestial object defs
 - Cache utilities which might be called through threads
+- Improve billboarding rotation
+- Support custom parameters for the object holder settle command